Change license from Unlicense to CC BY SA
I mainly changed the licence to CC BY SA to try and encourage others to make open their own design innovations based on this work. I have put a bit of thought into various types of licence, GNU GPL, CERN OHL but in the end decided to try and keep it pretty simple, not that those other licences aren't simple but more that CC is the licence brand leader.